fix(network): comprehensive security, correctness, and robustness pass
TCP fixes:
- H-1: Free recv buffer on timeout/close paths (memory leak)
- H-2: Free recv_exact buffer before rt_trap (memory leak)
- H-13: set_nonblocking now returns bool; errors checked at both call sites

HTTP fixes:
- H-3: Apply HTTP_MAX_BODY_SIZE guard to read_body_until_close_conn
- H-4: Cap read_line_conn buffer at 64 KB to prevent unbounded realloc
- H-5: Limit header count to 256 with drain loop for excess headers
- M-5: Guard negative atoll Content-Length before casting to size_t
- M-6: Overflow-safe hex chunk size parser (pre-multiply range check)
- M-13a: All HTTP convenience methods use strdup for method string
- RC-11: rt_http_post copies body with malloc+memcpy (not raw pointer)
- RC-12: Reject unrecognized URL schemes (ftp://, etc.) in parse_url
- RC-13: Drain all chunked trailer headers until empty line (RFC 7230)
- RC-14: Remove partial file on fwrite failure in rt_http_download

TLS fixes:
- CS-6: Default verify_cert=0 since X.509 validation is not implemented
- H-8: Sequence number overflow check; close on wrap (RFC 8446 §5.5)
- H-9: Constant-time Finished MAC comparison (ct_memcmp)
- H-10: Transcript buffer overflow aborts handshake instead of silently truncating
- M-13: Replace recursive rt_tls_recv retry with goto to prevent stack overflow

WebSocket fixes:
- CS-4: Use rt_crypto_random_bytes for masking key and WS key generation
- CS-5: Validate Sec-WebSocket-Accept header with inline SHA-1 + base64
- H-11: Implement FIN-bit fragmentation reassembly in recv and recv_bytes
- M-9: Reject server-masked frames with protocol error (RFC 6455 §5.1)

RestClient fixes:
- RC-1: Add rest_client_finalize; register via rt_obj_set_finalizer
- RC-2: Release previous last_response before overwriting in execute_request

RetryPolicy fixes:
- RC-5/RC-6: Check delay >= max and INT64_MAX/2 before multiplying
- RC-7: Add ±25% jitter to exponential backoff to prevent thundering herd

RateLimiter fixes:
- RC-9: Clamp tokens to 0 after subtraction to guard float precision underflow
- RC-10: Fall back to GetTickCount64 if QueryPerformanceFrequency fails

Documentation (docs/viperlib/network.md):
- D-1: WebSocket RecvFor returns null on timeout, not empty string
- D-2: Note Viper.Crypto.Tls low-level API and encryption-without-auth status
- D-3: Correct HTTPS TLS note; certificate chain validation not yet implemented
- D-4: Add TLS configuration guidance under HttpReq
- D-5/D-6: Document DNS error non-distinction and synchronous blocking behavior
- D-7: Document RateLimiter token precision and thread-safety requirements
- D-8: Clarify RetryPolicy attempt count semantics (0-based, N calls before exhaustion)
- D-9: Document Http.Post fixed Content-Type in method table and Limitations
- D-10: Add RestClient lifecycle note (LastResponse validity, GC behavior)

+++ src/lib/gui/src/font/vg_cache.c
@@ -4,7 +4,8 @@
 #include <string.h>
 
 // Monotonic tick incremented on every cache hit — used for LRU eviction.
-static uint32_t g_cache_tick = 0;
+// uint64_t prevents wrap-around after 4B+ hits in long-running applications.
+static uint64_t g_cache_tick = 0;
 
 //=============================================================================
 // Hash Function
@@ -154,8 +155,8 @@ static bool cache_resize(vg_glyph_cache_t *cache)
 
 static int compare_ticks(const void *a, const void *b)
 {
-    uint32_t ta = (*(const vg_cache_entry_t **)a)->access_tick;
-    uint32_t tb = (*(const vg_cache_entry_t **)b)->access_tick;
+    uint64_t ta = (*(const vg_cache_entry_t **)a)->access_tick;
+    uint64_t tb = (*(const vg_cache_entry_t **)b)->access_tick;
     return (ta < tb) ? -1 : (ta > tb) ? 1 : 0;
 }
